+++ Source/BugsnagConfiguration.m
@@ -40,7 +40,6 @@ static NSString *const kHeaderApiSentAt = @"Bugsnag-Sent-At";
 static NSString *const BSGApiKeyError = @"apiKey must be a 32-digit hexadecimal value.";
 static NSString *const BSGInitError = @"Init is unavailable.  Use [[BugsnagConfiguration alloc] initWithApiKey:] instead.";
 static const int BSGApiKeyLength = 32;
-NSString * const BSGConfigurationErrorDomain = @"com.Bugsnag.CocoaNotifier.Configuration";
 
 // User info persistence keys
 NSString * const kBugsnagUserKeychainAccount = @"BugsnagUserKeychainAccount";
@@ -104,15 +103,10 @@ NSString * const kBugsnagUserUserId = @"BugsnagUserUserId";
 /**
  * The designated initializer.
  */
--(instancetype)initWithApiKey:(NSString *)apiKey
-                        error:(NSError * _Nullable __autoreleasing * _Nullable)error
+- (instancetype _Nonnull)initWithApiKey:(NSString *_Nonnull)apiKey
 {
-    if (! [BugsnagConfiguration isValidApiKey:apiKey]) {
-        *error = [NSError errorWithDomain:BSGConfigurationErrorDomain
-                                     code:BSGConfigurationErrorInvalidApiKey
-                                 userInfo:@{NSLocalizedDescriptionKey : @"Invalid API key.  Should be a 32-digit hex string."}];
-        
-        return nil;
+    if (![BugsnagConfiguration isValidApiKey:apiKey]) {
+        bsg_log_err(@"Invalid configuration. apiKey should be a 32-character hexademical string, got \"%@\"", apiKey);
     }
     
     self = [super init];
@@ -355,6 +349,40 @@ NSString * const kBugsnagUserUserId = @"BugsnagUserUserId";
     self.breadcrumbs.capacity = capacity;
 }
 
+/**
+ * Specific types of breadcrumb should be recorded if either enabledBreadcrumbTypes
+ * is None, or contains the type.
+ *
+ * @param type The breadcrumb type to test
+ * @returns Whether to record the breadcrumb
+ */
+- (BOOL)shouldRecordBreadcrumbType:(BSGBreadcrumbType)type {
+    // enabledBreadcrumbTypes is BSGEnabledBreadcrumbTypeNone
+    if (!self.enabledBreadcrumbTypes) {
+        return YES;
+    }
+    
+    switch (type) {
+        case BSGBreadcrumbTypeManual:
+            return YES;
+        case BSGBreadcrumbTypeError :
+            return self.enabledBreadcrumbTypes & BSGEnabledBreadcrumbTypeError;
+        case BSGBreadcrumbTypeLog:
+            return self.enabledBreadcrumbTypes & BSGEnabledBreadcrumbTypeLog;
+        case BSGBreadcrumbTypeNavigation:
+            return self.enabledBreadcrumbTypes & BSGEnabledBreadcrumbTypeNavigation;
+        case BSGBreadcrumbTypeProcess:
+            return self.enabledBreadcrumbTypes & BSGEnabledBreadcrumbTypeProcess;
+        case BSGBreadcrumbTypeRequest:
+            return self.enabledBreadcrumbTypes & BSGEnabledBreadcrumbTypeRequest;
+        case BSGBreadcrumbTypeState:
+            return self.enabledBreadcrumbTypes & BSGEnabledBreadcrumbTypeState;
+        case BSGBreadcrumbTypeUser:
+            return self.enabledBreadcrumbTypes & BSGEnabledBreadcrumbTypeUser;
+    }
+    return NO;
+}
+
 // MARK: -
 
 @synthesize releaseStage = _releaseStage;
@@ -434,7 +462,7 @@ NSString * const kBugsnagUserUserId = @"BugsnagUserUserId";
     return self.autoTrackSessions;
 }
 
-// MARK: -
+// MARK: - enabledBreadcrumbTypes
 
 - (BSGEnabledBreadcrumbType)enabledBreadcrumbTypes {
     return self.breadcrumbs.enabledBreadcrumbTypes;